pyproject.toml
[tool.poetry]
name = "poetry-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "poetry_test"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
numpy = "^1.24.2"
kaggle = "^1.5.13"
markovchain = "^0.2.6"
markovify = "^0.9.4"
tweepy = "^4.14.0"
requests = "^2.30.0"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
discord = "^2.2.3"
python-dotenv = "^1.0.0"
nltk = "^3.8.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
